requirements.txt
anyio==4.12.1
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
distlib==0.4.0
filelock==3.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
nodeenv==1.10.0
packaging==26.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
ruff==0.15.0
urllib3==2.6.3
virtualenv==20.36.1
